+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # VAULT - Password Manager for the Decentralized Web
 
+### 1.3.0 (2021-05-27)
+* Update `skynet-js` to version `3.0.2`
+* `AES-CBC` algorithm replaced by `AES-GCM`
+* `localStorage` replaced by `sessionStorage`
+* Logo animation on saving
+* Code cleanup
 
 ### 1.2.0 (2020-12-05)
 * Code cleanup